Wire AdServices token migration through V4Migrator
Replace the inline migration in AttributionPoster.init with a proper
V4Migrator that fits the existing FileManagerMigrator pattern:

- Bump DataStoreVersion to v5.
- New V4Migrator (v4 → v5) reads the legacy user-specific
  AdServicesTokenStorage, writes it back at the new app-specific
  location, deletes the legacy file. Runs once via the version-keyed
  migration chain Storage.migrateData() already calls.
- LegacyUserScopedAdServicesTokenStorage moves out of CacheKeys.swift
  and lives next to the migrator, matching V3Migrator's pattern with
  LegacyLatestRedeemResponse.

The memCache delete-collision concern from the old inline approach
doesn't apply here: migration runs once at Storage init, before any
other code reads AdServicesTokenStorage, so deleting the legacy
entry doesn't evict a "just-written" memCache entry that any caller
depends on — the next read just falls through to the new app-specific
disk file.
